Feature/#640 (#651)
* Double zoom delivery - feature/#640

Changes not related to double zoom:
* CacheManager: new ITileSource member instead of MapTileProviderBase

Light changes:
* org.osmdroid.views.MapController: plus unrelated `zoomToSpan(int latSpanE6, int lonSpanE6)` rewriting
* org.osmdroid.google.wrapper.MapView
* MinimapOverlay
* MyLocationNewOverlay
* OpenStreetMapConstants: new tag in order to get/put the double zoom in prefs
* OsmdroidMapWrapper
* OsmPath
* Polyline
* Projection
* SampleOsmPath
* SimpleFastPointOverlay
* StarterMapFragment: using new tag in order to get/put the double zoom in prefs
* ZoomEvent

Tricky changes:
* IMapController: deprecated 3 integer zoom methods; created 3 double zoom methods instead
* IMapView: deprecated 1 integer zoom method; created 1 double zoom methods instead
* org.osmdroid.google.wrapper.MapController: plus unrelated `zoomToSpan(int latSpanE6, int lonSpanE6)` rewriting
* MapTileProviderBase
* org.osmdroid.views.MapView
* PathProjection: to be double-checked
* TileLooper: now we need to make the distinction between two tile sizes - for input it's basically 256, for output it's between 256 (integer zoom) and 511.99999 (almost the next integer zoom)
* TilesOverlay: handled the distinction between the two tile sizes, and the rounding issues
* org.osmdroid.util.TileSystem: deprecated some integer zoom methods, created the corresponding double zoom methods (among them `getFactor`, which computes real powers of 2 instead of << and >>), had to copy some code from microsoft.mappoint.TileSystem as we needed different tile sizes

* Double zoom delivery - feature/#640

Bug fix for @secaw

Changes not related to double zoom:
* MapTileProviderArray: removed unrelated duplicate code

Light changes:
* MinimapOverlay: rewrote call to `super.drawTiles`
* TileSystem: created `Rect PixelXYToTileXY` (`Rect` instead of `Point` as parameter)

Tricky changes:
* MapTileProviderBase: impacted the TileLooper refactoring; fixed an unrelated bug about LRUCache `remove`
* TileLooper: refactoring due to the confusion with tile sizes and viewports; fixed the standard "modulo+Pacific" bug
* TilesOverlay: impacted the TileLooper refactoring; fixed the standard "modulo+Pacific" bug

* Cosmetic changes - actually adding `@since` javadoc comments on new methods
* CacheManager
* IMapController
* IMapView
* org.osmdroid.views.MapController
* org.osmdroid.google.wrapper.MapController
* org.osmdroid.views.MapView
* org.osmdroid.google.wrapper.MapView
* Projection
* TilesOverlay
* org.osmdroid.util.TileSystem
* ZoomEvent

* Merge conflicts clean up

* Bug fix and optimization:
* MapTileModuleProviderBase: fixed the status for approximated tiles in method `TileLoader.run`
* MapTileProviderArray: added method `isDowngradedMode`, rewrote `getMapTile` accordingly
* MapTileProviderBasic: added member `mNetworkAvailabilityCheck`, added method `isDowngradedMode` that uses it
* MapTileRequestState: optimized queue

* Bug fix and optimization:
* MapTileModuleProviderBase: "oops" fix

* Bug fix and optimization:
* MapTileProviderArray: removed Log.e debug lines

* Bug fix:
* StarterMapActivity: the display is refreshed when switching from offline to online

* Concurrent branches copy:
* ExpirableBitmapDrawable
* IFilesystemCache
* MapTileApproximater
* SqliteArchiveTileWriter
* SqlTileWriter
* TileWriter

* Merge conflict solving and code cleaning:
* MapTileApproximater - removed Log.e call

* Merge conflict solving and code cleaning:
* MapTileApproximater - cleaned presentation

* Answer to failing checks: (#652)

* Bug445Caching: more verbose and more accurate version

+++ osmdroid-android/src/main/java/org/osmdroid/events/ZoomEvent.java
@@ -9,9 +9,9 @@ import org.osmdroid.views.MapView;
  */
 public class ZoomEvent implements MapEvent {
 	protected MapView source;
-	protected int zoomLevel;
+	protected double zoomLevel;
 
-	public ZoomEvent(final MapView source, final int zoomLevel) {
+	public ZoomEvent(final MapView source, final double zoomLevel) {
 		this.source = source;
 		this.zoomLevel = zoomLevel;
 	}
@@ -25,8 +25,9 @@ public class ZoomEvent implements MapEvent {
 
 	/*
 	 * Return the zoom level zoomed to.
+	 * Used to be an int, but is a double since 6.0
 	 */
-	public int getZoomLevel() {
+	public double getZoomLevel() {
 		return zoomLevel;
 	}
